Replace strip-indent with vendored function (#1)
The strip-indent package doesn't use any kind of transpilation,
but it does use arrow functions. In most environments dependencies
are not transpiled, so using this dependency will lead to issues.
While they're not hard to solve, it is probably a good idea to
just replace the module here, as there doesn't seem to be
any interest upstream to transpile the package:

https://github.com/sindresorhus/strip-indent/issues/1

I have not added a separate file since there's no bundling or
transpilation process in place here either. I hope that's
fine :smile: